Feat : Ram Func Fixed

+++ include/process.h
@@ -22,7 +22,7 @@ class Process {
   // TODO: Declare any necessary private members
  private:
  int PID_;
- long int ram_;
+ int ram_;
 };
 
 #endif
